불필요한 변수/ 메서드 제거, HighScoreScreen 수정. 현재로서는 점수-생존 시간 순으로 출력함.

+++ src/engine/Core.java
@@ -189,7 +189,7 @@ public final class Core {
                             gameState.getHp(),
                             gameState.getBulletsShot(),
                             gameState.getShipsDestroyed(),
-                            (int) gameState.getSurvivalTime());
+                            gameState.getSurvivalTime());
 
                     } while (gameState.getHp() > 0
                         && gameState.getLevel() <= NUM_LEVELS);
